backend/requirement.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
email-validator

# Database - MySQL
pymysql==1.1.0
sqlalchemy==2.0.23
cryptography==41.0.7  # Required for PyMySQL

# Data Processing
pyspark==3.5.0
pandas==2.1.3
numpy==1.26.2


# Environment and Configuration
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Async Support
asyncio==3.4.3

# Optional: Redis (if you want to add caching later)
# redis==5.0.1

# Optional: Web Scraping
# selenium==4.15.2
# beautifulsoup4==4.12.2
# requests==2.31.0

# Development Tools (optional)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0

# JWT Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# OAuth 2.0 (Google OAuth)
authlib==1.3.0
httpx==0.27.0
itsdangerous==2.1.2  # Required for SessionMiddleware

#Cloud Storage
boto3
